backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
razorpay==2.0.1
requests==2.32.3
pandas==2.2.2
cryptography==42.0.8
bcrypt==3.2.2
python-jwt==4.1.0
watchfiles==0.21.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.31.0
openpyxl==3.1.2
